+++ model/api.ts
@@ -36,13 +36,11 @@ export function api_updateGroupData(postData: SdImageGroup) {
   );
 }
 
-export function api_upsertStudy(postData: SdImageStudyDef) {
-  return axios.post<
-    SdImageStudyDef[],
-    AxiosResponse<SdImageStudyDef[]>,
-    SdImageStudyDef
-  >(`/api/studies`, postData as any);
-}
+export const api_upsertStudy = (postData) =>
+  simplePost<SdImageStudyDef, SdImageStudyDef[]>(`/api/studies`, postData);
+
+export const api_deleteStudy = (studyDef) =>
+  simpleDelete<SdImageStudyDef, any>(`/api/studies`, studyDef);
 
 export async function api_getStudy(id: string) {
   const url = getAbsUrl(`/api/studies/${id}`);
@@ -51,4 +49,22 @@ export async function api_getStudy(id: string) {
   const data = res.data;
 
   return data;
+}
+
+async function simplePost<TPostData, TResData>(url: string, data: any) {
+  const res = await axios.post<TResData, AxiosResponse<TResData>, TPostData>(
+    url,
+    data as any
+  );
+
+  return res.data;
+}
+
+async function simpleDelete<TDelData, TResData = any>(
+  url: string,
+  data: TDelData
+) {
+  const res = await axios.delete<TResData>(url, { data });
+
+  return res.data;
 }