+++ pages/api/studies/index.ts
@@ -1,4 +1,4 @@
-import { db_upsertStudy } from "../../../libs/db/studies";
+import { db_deleteStudy, db_upsertStudy } from "../../../libs/db/studies";
 
 import type { SdImageStudyDef } from "../../../libs/shared-types/src";
 
@@ -10,6 +10,18 @@ export default async function handler(req, res) {
     const studyData = req.body as SdImageStudyDef;
 
     const data = await db_upsertStudy(studyData);
+    res.status(200).json(data);
+    return;
+  }
+
+  // add a delete handler here
+  if (req.method === "DELETE") {
+    // delete the group with the body data
+
+    // get group from body
+    const studyData = req.body as SdImageStudyDef;
+
+    const data = await db_deleteStudy(studyData.id);
     res.status(200).json(data);
     return;
   }